Proper Noun Phrase Extraction
I've expanded the capabilities of the Proper Noun extractor and reduced
the waste from tagging the same word twice.  The tagger now runs on the
word list instead of every tweet, and can process much more information
without slowing the whole thing down.  With that said, the tagger is
still really slow.