New Event Class and data serialization
The program now uses the event class to create an event to store
actors, awards, and top tweeters.  The program also now guesses as to
what awards are (finds most of them), and who actors are.  All the
winners are in fact included in the actor list.  I also now store the
output of some of the functions into txt files so you can see what is
being generated.